Refactor code and remove unnecessary files.
