+++ public/index.html
@@ -29,11 +29,11 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-    <title>React App</title>
+    <title>Forex Information App</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
-    <div id="root"></div>
+    <div id="root" className="bg-slate-400 scroll-smooth"></div>
     <!--
       This HTML file is a template.
       If you open it directly in the browser, you will see an empty page.
